Adjust the edge cuts based on the space available on the phone
Rev v1.0.6 of the phone has a plastic bracket on the mid-frame which
joins the WiFi tab holder section to the section above the smartcard
slot. This piece bridges over the testi points and has holes where the
test points are located.

The intent of the edge cut design in this commit is to snake around the
plastic mid-frame that bridges between the two sections as was
described. If successful, this will neatly fit in the space remaining on
a v1.0.6 phone.

Also begin routing in this allotted space which tries to avoid the
plastic obstructions. Since avoiding the plastic brakets requires the
use of non-45 degree angles, this commit requires adjusting the end
points of the traces by hand with the use of some simple trig
calculations along with having to join some tracks with fillets in order
to accomplish the desired arc radii.

Tear drops are used in this commit thanks to NilujePerchut's handy
script (on commit 41c055d of the kicad_scripts repo).