Route the remaining portion of the FPC section and add the rigid section
The rigid section has a mounting hole which lines up with the center
screw that holds the M.2 cover in place. This will allow the board to be
mounted to the back of the phone.

Extra flex length was given to ensure enough distance to reach the
position where the rigid section will be mounted. The drawings on the
User.1 layer shows where the plastic obstacles are as well as the
location of the M1.4 threaded hole. There is approximately 5.4mm of
additional slack in the length of the flex section so that it can be
tucked inside the volume it runs along.

The rigid section of the board will sit between the M.2 modules, on top
of the ~10.5mm plastic section that separates the two. The rigid section
has a 7mm and should close to the center of the plastic divider.