Make the board 4 layers and move the flex copper to the inner layers
The 0402 caps and ferrite beads and their traces have to remain on the
F.Cu layer.

This commit was accomplished by having to manually edit the .kicad_pcb
text file in order to move the previous F.Cu & B.Cu layers to becoming
In1.Cu & In2.Cu.

Also add notes regarding where the rigid section is, where the flex
section is, and that the area with the FPC connector should have a
kapton stiffener placed there. These notes are found on the
User.Comments layer.