Export the manufacturing files (gerbers and PDF)
Add fab notes on the User.Comments layer. Include a general stackup
guide image in the PDF fabrication drawing.

Re-centered the whole board on the canvas. The vias on the flex section
have to be set as "buried" vias going from In1.Cu to In2.Cu in order to
get the correct gerber export.